+++ internal/goscrobble/album.go
@@ -3,6 +3,7 @@ package goscrobble
 import (
 	"database/sql"
 	"errors"
+	"fmt"
 	"log"
 )
 
@@ -12,41 +13,47 @@ type Album struct {
 	Desc          sql.NullString `json:"desc"`
 	Img           sql.NullString `json:"img"`
 	MusicBrainzID sql.NullString `json:"mbid"`
+	SpotifyID     sql.NullString `json:"spotify_id"`
 }
 
 // insertAlbum - This will return if it exists or create it based on MBID > Name
-func insertAlbum(name string, mbid string, artists []string, tx *sql.Tx) (Album, error) {
+func insertAlbum(name string, mbid string, spotifyId string, artists []string, tx *sql.Tx) (Album, error) {
 	album := Album{}
 
+	// Try locate our album
 	if mbid != "" {
 		album = fetchAlbum("mbid", mbid, tx)
-		if album.Uuid == "" {
-			err := insertNewAlbum(name, mbid, tx)
-			if err != nil {
-				log.Printf("Error inserting album via MBID %s  %+v", name, err)
-				return album, errors.New("Failed to insert album")
-			}
+	} else if spotifyId != "" {
+		album = fetchAlbum("spotify_id", spotifyId, tx)
+	}
 
-			album = fetchAlbum("mbid", mbid, tx)
-			err = album.linkAlbumToArtists(artists, tx)
-			if err != nil {
-				return album, err
-			}
-		}
-	} else {
+	// If it didn't match above, lookup by name
+	if album.Uuid == "" {
 		album = fetchAlbum("name", name, tx)
-		if album.Uuid == "" {
-			err := insertNewAlbum(name, mbid, tx)
-			if err != nil {
-				log.Printf("Error inserting album via Name %s %+v", name, err)
-				return album, errors.New("Failed to insert album")
-			}
+	}
 
+	// If we can't find it. Lets add it!
+	if album.Uuid == "" {
+		err := insertNewAlbum(name, mbid, spotifyId, tx)
+		if err != nil {
+			return album, errors.New("Failed to insert album")
+		}
+
+		// Fetch the recently inserted album to get the UUID
+		if mbid != "" {
+			album = fetchAlbum("mbid", mbid, tx)
+		} else if spotifyId != "" {
+			album = fetchAlbum("spotify_id", spotifyId, tx)
+		}
+
+		if album.Uuid == "" {
 			album = fetchAlbum("name", name, tx)
-			err = album.linkAlbumToArtists(artists, tx)
-			if err != nil {
-				return album, err
-			}
+		}
+
+		// Try linkem up
+		err = album.linkAlbumToArtists(artists, tx)
+		if err != nil {
+			return album, errors.New("Unable to link albums!")
 		}
 	}
 
@@ -72,9 +79,9 @@ func fetchAlbum(col string, val string, tx *sql.Tx) Album {
 	return album
 }
 
-func insertNewAlbum(name string, mbid string, tx *sql.Tx) error {
-	_, err := tx.Exec("INSERT INTO `albums` (`uuid`, `name`, `mbid`) "+
-		"VALUES (UUID_TO_BIN(UUID(), true),?,?)", name, mbid)
+func insertNewAlbum(name string, mbid string, spotifyId string, tx *sql.Tx) error {
+	_, err := tx.Exec("INSERT INTO `albums` (`uuid`, `name`, `mbid`, `spotify_id`) "+
+		"VALUES (UUID_TO_BIN(UUID(), true),?,?,?)", name, mbid, spotifyId)
 
 	return err
 }
@@ -85,9 +92,16 @@ func (album *Album) linkAlbumToArtists(artists []string, tx *sql.Tx) error {
 		_, err = tx.Exec("INSERT INTO `album_artist` (`album`, `artist`) "+
 			"VALUES (UUID_TO_BIN(?, true), UUID_TO_BIN(?, true))", album.Uuid, artist)
 		if err != nil {
+			fmt.Println(err)
 			return err
 		}
 	}
+
+	return err
+}
+
+func updateAlbum(uuid string, col string, val string, tx *sql.Tx) error {
+	_, err := tx.Exec("UPDATE `albums` SET `"+col+"` = ? WHERE `uuid` = UUID_TO_BIN(?,true)", val, uuid)
 
 	return err
 }