+++ internal/goscrobble/artist.go
@@ -12,34 +12,43 @@ type Artist struct {
 	Desc          sql.NullString `json:"desc"`
 	Img           sql.NullString `json:"img"`
 	MusicBrainzID sql.NullString `json:"mbid"`
+	SpotifyID     sql.NullString `json:"spotify_id"`
 }
 
 // insertArtist - This will return if it exists or create it based on MBID > Name
-func insertArtist(name string, mbid string, tx *sql.Tx) (Artist, error) {
+func insertArtist(name string, mbid string, spotifyId string, tx *sql.Tx) (Artist, error) {
 	artist := Artist{}
 
+	// Try locate our artist
 	if mbid != "" {
 		artist = fetchArtist("mbid", mbid, tx)
-		if artist.Uuid == "" {
-			err := insertNewArtist(name, mbid, tx)
-			if err != nil {
-				log.Printf("Error inserting artist via MBID %s  %+v", name, err)
-				return artist, errors.New("Failed to insert artist")
-			}
+	} else if spotifyId != "" {
+		artist = fetchArtist("spotify_id", spotifyId, tx)
+	}
 
-			artist = fetchArtist("mbid", mbid, tx)
-		}
-	} else {
+	// If it didn't match above, lookup by name
+	if artist.Uuid == "" {
 		artist = fetchArtist("name", name, tx)
-		if artist.Uuid == "" {
-			err := insertNewArtist(name, mbid, tx)
-			if err != nil {
-				log.Printf("Error inserting artist via Name %s %+v", name, err)
-				return artist, errors.New("Failed to insert artist")
-			}
+	}
 
-			artist = fetchArtist("name", name, tx)
+	// If we can't find it. Lets add it!
+	if artist.Uuid == "" {
+		err := insertNewArtist(name, mbid, spotifyId, tx)
+		if err != nil {
+			log.Printf("Error inserting artist: %+v", err)
+			return artist, errors.New("Failed to insert artist")
 		}
+	}
+
+	// Fetch the recently inserted artist to get the UUID
+	if mbid != "" {
+		artist = fetchArtist("mbid", mbid, tx)
+	} else if spotifyId != "" {
+		artist = fetchArtist("spotify_id", spotifyId, tx)
+	}
+
+	if artist.Uuid == "" {
+		artist = fetchArtist("name", name, tx)
 	}
 
 	if artist.Uuid == "" {
@@ -64,9 +73,15 @@ func fetchArtist(col string, val string, tx *sql.Tx) Artist {
 	return artist
 }
 
-func insertNewArtist(name string, mbid string, tx *sql.Tx) error {
-	_, err := tx.Exec("INSERT INTO `artists` (`uuid`, `name`, `mbid`) "+
-		"VALUES (UUID_TO_BIN(UUID(), true),?,?)", name, mbid)
+func insertNewArtist(name string, mbid string, spotifyId string, tx *sql.Tx) error {
+	_, err := tx.Exec("INSERT INTO `artists` (`uuid`, `name`, `mbid`, `spotify_id`) "+
+		"VALUES (UUID_TO_BIN(UUID(), true),?,?,?)", name, mbid, spotifyId)
+
+	return err
+}
+
+func updateArtist(uuid string, col string, val string, tx *sql.Tx) error {
+	_, err := tx.Exec("UPDATE `artists` SET `"+col+"` = ? WHERE `uuid` = UUID_TO_BIN(?,true)", val, uuid)
 
 	return err
 }