+++ internal/goscrobble/ingress_jellyfin.go
@@ -6,12 +6,13 @@ import (
 	"fmt"
 	"log"
 	"net"
+	"time"
 )
 
-// ParseJellyfinInput - Transform API data into a common struct
+// ParseJellyfinInput - Transform API data into a common struct. Uses MusicBrainzID primarily
 func ParseJellyfinInput(userUUID string, data map[string]interface{}, ip net.IP, tx *sql.Tx) error {
 	// Debugging
-	fmt.Printf("%+v", data)
+	// fmt.Printf("%+v", data)
 
 	if data["ItemType"] != "Audio" {
 		return errors.New("Media type not audio")
@@ -31,7 +32,7 @@ func ParseJellyfinInput(userUUID string, data map[string]interface{}, ip net.IP,
 	}
 
 	// Insert artist if not exist
-	artist, err := insertArtist(fmt.Sprintf("%s", data["Artist"]), fmt.Sprintf("%s", data["Provider_musicbrainzartist"]), tx)
+	artist, err := insertArtist(fmt.Sprintf("%s", data["Artist"]), fmt.Sprintf("%s", data["Provider_musicbrainzartist"]), "", tx)
 	if err != nil {
 		log.Printf("%+v", err)
 		return errors.New("Failed to map artist")
@@ -39,29 +40,28 @@ func ParseJellyfinInput(userUUID string, data map[string]interface{}, ip net.IP,
 
 	// Insert album if not exist
 	artists := []string{artist.Uuid}
-	album, err := insertAlbum(fmt.Sprintf("%s", data["Album"]), fmt.Sprintf("%s", data["Provider_musicbrainzalbum"]), artists, tx)
+	album, err := insertAlbum(fmt.Sprintf("%s", data["Album"]), fmt.Sprintf("%s", data["Provider_musicbrainzalbum"]), "", artists, tx)
 	if err != nil {
 		log.Printf("%+v", err)
 		return errors.New("Failed to map album")
 	}
 
-	// Insert album if not exist
-	track, err := insertTrack(fmt.Sprintf("%s", data["Name"]), fmt.Sprintf("%s", data["Provider_musicbrainztrack"]), album.Uuid, artists, tx)
+	// Insert track if not exist
+	length := timestampToSeconds(fmt.Sprintf("%s", data["RunTime"]))
+	track, err := insertTrack(fmt.Sprintf("%s", data["Name"]), length, fmt.Sprintf("%s", data["Provider_musicbrainztrack"]), "", album.Uuid, artists, tx)
 	if err != nil {
 		log.Printf("%+v", err)
 		return errors.New("Failed to map track")
 	}
 
-	// Insert album if not exist
-	err = insertScrobble(userUUID, track.Uuid, "jellyfin", ip, tx)
+	// Insert scrobble if not exist
+	timestamp := time.Now()
+	fmt.Println(timestamp)
+	err = insertScrobble(userUUID, track.Uuid, "jellyfin", timestamp, ip, tx)
 	if err != nil {
 		log.Printf("%+v", err)
 		return errors.New("Failed to map track")
 	}
-
-	_ = album
-	_ = artist
-	_ = track
 
 	// Insert track if not exist
 	return nil