+++ migrations/10_tracklength.up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `tracks` ADD COLUMN `length` INT NOT NULL DEFAULT 0;